Cache blobs
Profiling pointed to this as the next biggest sink of CPU time. This
shaves 5 seconds off the runtime of scc on the linux kernel.

This is a bad implementation, because we now keep all these blob objects
alive until the end of the program. I guess we want to tie these to
inode refcounts, so that once we no longer need an inode we can
garbage-collect all the attributes and blobs that correspond to it.